requirements.txt
fastapi== 0.95.0
uvicorn[standard]== 0.21.1
python-multipart==0.0.6
pydantic==1.10.7
pydantic[email]
pymongo==4.3.3
python-decouple==3.8
PyJWT==2.6.0
